Fixing dot for status effect (#211)
* WIP working DOT but not die of DOT is possible

* WIP working on affector data

* added affector to DOT

* added additional damage to fire effect

+++ CSharpSourceCode/Battle/StatusEffects/StatusEffect.cs
@@ -34,7 +34,7 @@ namespace TOW_Core.Battle.StatusEffects
         public EffectType Type { get; set; }
 
         [XmlIgnore]
-        public int CurrentDuration { get; set; }
+        public float CurrentDuration { get; set; }
         [XmlIgnore]
         public Agent Affector { get; set; }
